# Enh/svd with floats