--- a/src/SymbolicWedderburn.jl
+++ b/src/SymbolicWedderburn.jl
@@ -23,5 +23,6 @@ include("dixon.jl")
 
 include("actions.jl")
 include("projections.jl")
+include("sa_basis.jl")
 
 end # module