… Project.toml
… examples/ex_C2_linear.jl
… examples/ex_C4.jl
… examples/ex_motzkin.jl
… test/characters.jl
… test/dixon.jl
… test/projections.jl
… test/runtests.jl
… test/sa_basis.jl