changes here include:

* add `image_basis` that dispatches either to `row_echelon_form` (e.g. for exact types), or uses `svd` to guess rank and finds orthonormal basis (for `BlasFloats`/`BlasComplex`).
* rework interface to `symmetry_adapted_basis`: we have two type arguments in the signature: `symmetry_adapted_basis(::Type{T}, G, basis, action,  ::Type{S}=Rational{Int})`
  - `T` determines the type of coefficients for the returned basis; e.g. we dispatch to the fast BLAS routines for `image_basis`. The rationale is that if someone wants `Float64` basis, then maybe there is little point of doing everything symbolically. In particular `T<:Real` means that we return real basis.
  - `S` controls the coefficients over which we compute characters. For some cases one needs to go to `Rational{BigInt}` (e.g. cyclic groups with cycle longer than ~17)
* I bumped version to `v0.1.1`, but I think `v0.2` is justified, as by default (without specifying either `S` or `T`) we compute complex basis now.
* in the tests we run over all small finite groups up to order 30, as `symmetry_adapted_basis` got much faster (though not necessarily in this pull request)