@kalmarek 2021-08-18
@blegat this seems ready from my point of view. I'm leaning more to `v"0.2.0"` now, versions cost penny a pair ;)