@kalmarek 2021-08-25
Recent improvements (mostly rref)
```julia
include("test/dixon.jl") # second run
include("test/sa_basis") # second run
...
now:
```
# dixon
  4.641992 seconds (20.00 M allocations: 1.703 GiB, 11.47% gc time, 24.51% compilation time)
# sa_basis
  6.780373 seconds (126.39 M allocations: 5.448 GiB, 35.00% gc time) # Rational{Int}
  1.286905 seconds (7.61 M allocations: 745.589 MiB, 27.15% gc time, 1.30% compilation time)
```
before:
```
# dixon
  7.228704 seconds (30.31 M allocations: 3.836 GiB, 14.74% gc time, 15.64% compilation time)
# sa_basis
  8.197061 seconds (135.25 M allocations: 6.722 GiB, 32.53% gc time)
  1.562904 seconds (11.03 M allocations: 1.437 GiB, 18.63% gc time, 0.75% compilation time)
```

also with those improvements I can run a `1777×1777` decomposition in ~4s (used to be ~90)
with `Rational{Int}`-valued `Characters`. With `Float64`-valued it is much slower (due to oversubscriptions of cores = BLAS), but `svd` takes the majority of time: 52s. Note that we also get orthogonal basis with Float64s, which is much more than just rref.

BUT: can we orthogonalize the returned rational basis faster than svd does?